Make message close method bit more bulletproof
Noe, it it possible to create Message object and perform close() on it,
even if show() wasn't called before. It have a bit more sense since it
should be harmless for closing possibly non existent message.

+++ fs_uae_wrapper/message.py
@@ -51,9 +51,10 @@ class Message(object):
 
     def close(self):
         """Terminate the process with gui"""
-        if self._process.is_alive():
-            self._process.terminate()
-        self._process.join()
+        if self._process:
+            if self._process.is_alive():
+                self._process.terminate()
+            self._process.join()
 
 
 def _spawn(msg):
